refactor(rendering): change RenderBox::size() to return references
- Change size() from Size to &Size (avoid unnecessary copies)
- Add size_mut() -> &mut Size for mutable access
- Update all RenderBox implementations (padding, center, flex, etc.)
- Simplify blanket impl: layout/paint/hit_test go through PipelineOwner
- Add child_count()/child_offset()/set_was_repaint_boundary() to RenderObject trait
- Remove mark_needs_layout/paint from RenderEntry/RenderNode (require element_id+tree)
- Fix type conversions (u16 -> usize) and imports
- Fix RenderSliver::geometry() ambiguity
- Library compiles successfully, tests need API updates

Architecture note: Blanket impl exists for type compatibility only.
Real layout/paint/hit_test must go through PipelineOwner which has
access to RenderTree and can create proper contexts with children.

+++ crates/flui_rendering/src/pipeline/owner.rs
@@ -339,11 +339,11 @@ impl PipelineOwner {
         let child_depth = parent_depth + 1;
 
         // Mark child as needing layout and paint
-        self.add_node_needing_layout(child_id.get(), child_depth);
-        self.add_node_needing_paint(child_id.get(), child_depth);
+        self.add_node_needing_layout(child_id.get(), child_depth as usize);
+        self.add_node_needing_paint(child_id.get(), child_depth as usize);
 
         // Mark parent as needing layout (child structure changed)
-        self.add_node_needing_layout(parent_id.get(), parent_depth);
+        self.add_node_needing_layout(parent_id.get(), parent_depth as usize);
 
         Some(child_id)
     }
